chore: validate avatar url in handler
